requirements.txt
faster-whisper>=1.1.0
sounddevice>=0.5.0
numpy>=1.24.0
pynput>=1.7.7
pystray>=0.19.5
Pillow>=10.0.0
pyperclip>=1.9.0
PySide6>=6.7.0
